\name{gtcorr-package}
\alias{gtcorr-package}
\alias{gtcorr}
\docType{package}
\title{
Calculate efficiencies of group testing algorithms with correlated
responses
}
\description{
This package provides functions to calculate the efficiencies (expected
tests per unit) of hierarchical and matrix group testing procedures.
Efficiencies can be calculated in the presence of correlated responses
under multiple arrangements of clusters.  Efficiencies can also be
evaluated in the presence of test error.

}
\author{
Sam Lendle <lendle@stat.berkeley.edu>
}
\references{
Samuel D. Lendle, Michael Hudgens, and Bahjat F. Qaqish, "Group Testing
for Case Identification with Correlated Responses" Submitted 2011.
Biometrics.
}
\keyword{ package }

